feat: move Config to config package
